# Conflicts:
#	fa-base/src/main/java/com/yidu/deposit/dao/CashInventoryDao.java
#	fa-base/src/main/java/com/yidu/deposit/domain/CashInventory.java
#	fa-base/src/main/java/com/yidu/deposit/paging/CashInventoryPaging.java
#	fa-base/src/main/java/com/yidu/deposit/service/CashInventoryService.java
#	fa-base/src/main/java/com/yidu/deposit/service/impl/CashInventoryServiceImpl.java
#	fa-web/src/main/java/com/yidu/deposit/controller/CashInventoryController.java
#	fa-web/src/main/resources/applicationContext.xml
#	fa-web/src/main/webapp/static/fund/js/calculateRevenue.js
#	fa-web/src/main/webapp/static/fund/js/cashInterest.js
#	fa-web/src/main/webapp/static/fund/js/feeCalculate.js
#	fa-web/src/main/webapp/static/fund/page/calculateRevenue.html

+++ fa-base/src/main/java/com/yidu/account/domain/Account.java
@@ -4,6 +4,7 @@ import com.yidu.utils.DateUtils;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -38,7 +39,7 @@ public class Account implements Serializable{
 	/**
 	 * 存款利率,年利率
 	 */
-	private double depositRate;
+	private BigDecimal depositRate;
 
 	/**
 	 * 开户银行
@@ -121,11 +122,11 @@ public class Account implements Serializable{
 		this.depositType = depositType;
 	}
 
-	public double getDepositRate() {
+	public BigDecimal getDepositRate() {
 		return depositRate;
 	}
 
-	public void setDepositRate(double depositRate) {
+	public void setDepositRate(BigDecimal depositRate) {
 		this.depositRate = depositRate;
 	}
